AI-Models/requirements.txt
fastapi
uvicorn
torchvision
timm
pillow
numpy
python-multipart
transformers
requests
pandas
scikit-learn
tensorflow
tensorflow_hub
diffusers
pydantic
accelerate
safetensors
peft
git+https://github.com/openai/CLIP.git